Readme review
Hi @breedx-nr - I did a Readme review and made some suggestions. I wanted to ensure that customers who might happen upon this felt like we're talking to them, so I changed the wording a little. Also, in the Support section, please add a link to the community section for this project. If there's not one, you might ask Melissa Klein where in the community to point people with questions.